Gjorde image / category controller helt ferdig, og fikset litt på enhetstestene

+++ BLL/Category/CategoryBLL.cs
@@ -10,7 +10,7 @@ namespace BLL.Category
 {
     public class CategoryBLL :  ICategoryLogic
     {
-
+        
         private ICategoryRepo repo;
 
         public CategoryBLL()
@@ -23,49 +23,35 @@ namespace BLL.Category
             this.repo = repo;
         }
 
-        //public List<CategoryModel> GetAllCategoryModels()
-        //{
-        //    return repo.GetAllCategoryModels();
-        //}
 
         public bool AddCategory(string Name)
         {
             return repo.AddCategory(Name);
         }
 
-        public bool DeleteCategory(int CategoryId)
+        public bool DeleteCategory(int CategoryId, int AdminId)
         {
-            //Category category = repo.GetCategory(CategoryId);
+            Nettbutikk.Model.Category category = repo.GetCategory(CategoryId);
 
-            //if (category == null)
-            //    return false;
+            if (category == null)
+                return false;
 
-            //if (!repo.AddOldCategory(category.Name, 1))//Get admin id from session
-            //    return false;
+            if (!repo.AddOldCategory(category.Name, AdminId))
+                return false;
 
             return repo.DeleteCategory(CategoryId);
         }
 
-        //public List<Category> GetAllCategories()
-        //{
-        //    return repo.GetAllCategories();
-        //}
-
-        //public Category GetCategory(int CategoryId)
-        //{
-        //    return repo.GetCategory(CategoryId);
-        //}
-
-        public bool UpdateCategory(int CategoryId, string Name)
+        public bool UpdateCategory(int CategoryId, string Name, int AdminId)
         {
 
-            //Category category = repo.GetCategory(CategoryId);
+            Nettbutikk.Model.Category category = repo.GetCategory(CategoryId);
 
-            //if (category == null)
-            //    return false;
+            if (category == null)
+                return false;
 
-            //if (!repo.AddOldCategory(category.Name, 1))//Get admin id from session
-            //    return false;
+            if (!repo.AddOldCategory(category.Name, AdminId))
+                return false;
 
             return repo.UpdateCategory(CategoryId, Name);
         }
@@ -75,19 +61,32 @@ namespace BLL.Category
             return repo.GetCategoryName(CategoryId);
         }
 
-        public List<CategoryModel> GetAllCategories()
-        {
-            return repo.GetAllCategories();
-        }
-
-        public CategoryModel GetCategory(int CategoryId)
-        {
-            return repo.GetCategory(CategoryId);
-        }
 
         public int FirstCategoryWithProducts()
         {
             return repo.FirstCategoryWithProducts();
         }
+
+        public Nettbutikk.Model.Category GetCategory(int CategoryId)
+        {
+            return repo.GetCategory(CategoryId);
+        }
+
+        public List<Nettbutikk.Model.Category> GetAllCategories()
+        {
+            return repo.GetAllCategories();
+        }
+
+
+        public CategoryModel GetCategoryModel(int CategoryId)
+        {
+            return repo.GetCategoryModel(CategoryId);
+        }
+
+        public List<CategoryModel> GetAllCategoryModels()
+        {
+            return repo.GetAllCategoryModels();
+        }
+
     }
 }
